pyproject.toml
[tool.poetry]
name = "raindroppy"
version = "0.0.5"
description = "API and terminal-based CLI for Raindrop.io bookmark manager"
authors = ["Peter Borocz <680893+PBorocz@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/PBorocz/raindroppy"
homepage = "https://github.com/PBorocz/raindroppy"
include = ["LICENSE",]
keywords = ["python"]
classifiers = [
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3',
]

packages = [{ include = "raindroppy" }]

[tool.poetry.scripts]
raindroppy_cli = "raindroppy:cli:run.main"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
requests-oauthlib = "^1.3.1"
python-dateutil = "^2.8.2"
jashin = ">=0.0.6"
fire = "^0.5.0"
tomli = "^2.0.1"
humanize = "^4.4.0"
pyfiglet = "^0.8.post1"
prompt-toolkit = "^3.0.36"
rich = "^13.1.0"

[tool.poetry.dev-dependencies]
PySnooper = "^1.1.1"
python-language-server = "^0.36.2"
pre-commit = "^2.21.0"
pytest = "7.2.1"
flake8 = "^6.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: marks tests as runnable all the time, locally and fast.",
    "integration: tests that run against a live Raindrop environment, slower.",
]

# When we're ready (from https://fpgmaas.github.io/cookiecutter-poetry/)
# [tool.mypy]
# files = ["example_project"]
# disallow_untyped_defs = "True"
# disallow_any_unimported = "True"
# no_implicit_optional = "True"
# check_untyped_defs = "True"
# warn_return_any = "True"
# warn_unused_ignores = "True"
# show_error_codes = "True"

[tool.ruff]
target-version = "py310"
line-length = 120
select = [
       # Which Ruff suites are we running?
       # (presented in order from the Ruff documentation page)
       "F",                    # Pyflakes
       "E", "W",               # Pycodestyle (Errors and Warnings respectively)
       "C90",                  # Mccabe
       "I",                    # Isort
       "D",                    # Pydocstyle
       # "UP",                 # pyupgrade
       # "N",                  # pep8-naming
       "YTT",                  # flake8-2020
       # "ANN",                # flake8-annotations
       # "S",                  # flake8-bandit
       # "BLE",                # flake8-blind-except
       # "FBT",                # flake8-boolean-trap
       # "B",                  # flake8-bugbear
       # "A",                  # flake8-builtins
       # "C4",                 # flake8-comprehensions
       # "T10",                # flake8-debugger
       # "EM",                 # flake8-errmsg
       # "ISC",                # flake8-implicit-str-concat
       # "ICN",                # flake8-import-conventions
       # "T20",                # flake8-print
       # "PT",                 # flake8-pytest-style
       # "Q",                  # flake8-quotes
       # "RET",                # flake8-return
       # "SIM",                # flake8-simplify
       # "TID",                # flake8-tidy-imports
       # "ARG",                # flake8-unused-arguments
       # "DTZ",                # flake8-datetimez
       # "ERA",                # eradicate
       # "PD",                 # pandas-vet
       # "PGH",                # pygrep-hooks
       # "PLC, PLE, PLR, PLW", # Pylint
       # "PIE",                # flake8-pie
       "COM",                  # flake8-commas
       # "RUF",                # Ruff-specific rules
       ]
ignore = ["D213", "E402"]

[tool.ruff.pydocstyle]
# For more info, see:
# https://github.com/charliermarsh/ruff#does-ruff-support-numpy--or-google-style-docstrings
convention = "google"

[tool.ruff.mccabe]
max-complexity = 13

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools"]
build-backend = "poetry.core.masonry.api"

# Based on PublishingPythonPackages, Hillard 2022
# [build-system]
# requires = ["setuptools", "wheel"]
# build-backend = "setuptools.build_meta"

# [project]
# name = "RaindropPy"
# authors = [
#     {name = "Peter Borocz", email = "peter.borocz@gmail.com"},
# ]
# license = {text = "MIT"}
# description = "API & Command-line-interface for the Raindrop Bookmarking Service"
# readme = "README.md"
# dynamic = ["version"]

# [console.scripts]
# raindroppy-cli = "raindroppy:cli:run.main"

# [tool.setuptools.dynamic]
# version = {attr = "raindroppy.__version__"}

# requires-python = ">=3.9"
# classifiers = ["Framework :: Django", "Programming Language :: Python :: 3"]
# dependencies = ["requests", 'importlib-metadata; python_version<"3.8"' ]
